src/agents/AnalysisAgent.ts
import { BaseAgent } from '../core/BaseAgent';
import { LearningDistributor } from '../core/learning/LearningDistributor';
import { AgentCoordinator } from '../core/coordinator/AgentCoordinator';
import { Pattern, AnalysisResult, Confidence } from '../types';

export class AnalysisAgent extends BaseAgent {
    private learningDistributor: LearningDistributor;
    private coordinator: AgentCoordinator;
    private patternCache: Map<string, Pattern>;
    private confidenceThreshold: number;

    constructor(
        coordinator: AgentCoordinator,
        learningDistributor: LearningDistributor,
        config: AnalysisAgentConfig
    ) {
        super('analysis');
        this.coordinator = coordinator;
        this.learningDistributor = learningDistributor;
        this.patternCache = new Map();
        this.confidenceThreshold = config.confidenceThreshold ?? 0.75;
        
        this.initializePatternLearning();
    }

    private async initializePatternLearning(): Promise<void> {
        // Subscribe to pattern updates from the LearningDistributor
        this.learningDistributor.subscribe('pattern-update', (pattern: Pattern) => {
            this.updatePattern(pattern);
        });

        // Register with coordinator for inter-agent communication
        this.coordinator.register(this, ['assessment-result', 'pattern-request']);
    }

    private updatePattern(pattern: Pattern): void {
        this.patternCache.set(pattern.id, {
            ...pattern,
            lastUpdated: Date.now()
        });
    }

    public async analyze(data: any): Promise<AnalysisResult> {
        const patterns = await this.identifyRelevantPatterns(data);
        const assessmentResults = await this.coordinator.request('assessment-result', { dataId: data.id });
        
        const analysis: AnalysisResult = {
            id: data.id,
            timestamp: Date.now(),
            patterns: patterns,
            confidence: this.calculateConfidence(patterns, assessmentResults),
            insights: await this.generateInsights(data, patterns, assessmentResults)
        };

        await this.learningDistributor.distribute({
            type: 'analysis-complete',
            data: analysis,
            source: this.id
        });

        return analysis;
    }

    private async identifyRelevantPatterns(data: any): Promise<Pattern[]> {
        const relevantPatterns: Pattern[] = [];
        
        for (const [_, pattern] of this.patternCache) {
            const match = await this.evaluatePatternMatch(pattern, data);
            if (match.confidence > this.confidenceThreshold) {
                relevantPatterns.push(pattern);
            }
        }

        return relevantPatterns;
    }

    private async evaluatePatternMatch(pattern: Pattern, data: any): Promise<Confidence> {
        // Implement pattern matching logic using your established validation approach
        const matchScore = await this.calculatePatternMatchScore(pattern, data);
        
        return {
            score: matchScore,
            factors: {
                patternRelevance: this.calculatePatternRelevance(pattern, data),
                dataCompleteness: this.assessDataCompleteness(data),
                historicalAccuracy: await this.getHistoricalAccuracy(pattern.id)
            }
        };
    }

    private async generateInsights(
        data: any,
        patterns: Pattern[],
        assessmentResults: any
    ): Promise<string[]> {
        const insights: string[] = [];
        
        // Combine pattern knowledge with assessment results
        for (const pattern of patterns) {
            const insight = await this.synthesizeInsight(pattern, data, assessmentResults);
            if (insight) {
                insights.push(insight);
            }
        }

        return insights;
    }

    private calculateConfidence(patterns: Pattern[], assessmentResults: any): Confidence {
        const baseScore = patterns.reduce((acc, pattern) => 
            acc + (pattern.confidence ?? 0), 0) / patterns.length;
            
        return {
            score: baseScore,
            factors: {
                patternCount: patterns.length,
                assessmentConfidence: assessmentResults.confidence?.score ?? 0,
                patternConsistency: this.calculatePatternConsistency(patterns)
            }
        };
    }

    // Helper methods for confidence calculation and pattern matching
    private async calculatePatternMatchScore(pattern: Pattern, data: any): Promise<number> {
        // Implement your specific pattern matching logic here
        return 0.0; // Placeholder
    }

    private calculatePatternRelevance(pattern: Pattern, data: any): number {
        // Implement relevance calculation
        return 0.0; // Placeholder
    }

    private assessDataCompleteness(data: any): number {
        // Implement data completeness assessment
        return 0.0; // Placeholder
    }

    private async getHistoricalAccuracy(patternId: string): Promise<number> {
        // Implement historical accuracy retrieval
        return 0.0; // Placeholder
    }

    private calculatePatternConsistency(patterns: Pattern[]): number {
        // Implement pattern consistency calculation
        return 0.0; // Placeholder
    }

    private async synthesizeInsight(
        pattern: Pattern,
        data: any,
        assessmentResults: any
    ): Promise<string | null> {
        // Implement insight synthesis logic
        return null; // Placeholder
    }
}